+++ CHANGES.rst
@@ -1,6 +1,7 @@
 CHANGES
 -------
-- 2021-09-30 0.20 unittest
+- 2021-10-04 0.21 refactoring: KanaText, KanaValue.
+- 2021-09-30 0.20 refactoring: unittest.
 - 2021-09-30 0.19 handle homonymous functions by each name.
 - 2021-09-28 0.18 command 'sphinx-kana-genindex'
 - 2021-09-27 0.17 released on pypi.